feat(perf-detector-threshold-configuration) Added new thresholds and lowered existing thresholds.
