+++ tests/sentry/api/endpoints/test_project_performance_issue_settings.py
@@ -92,6 +92,7 @@ class ProjectPerformanceIssueSettingsTest(APITestCase):
                 "performance.issues.uncompressed_asset.size_threshold": 200000,
                 "performance.issues.consecutive_db.min_time_saved_threshold": 300,
                 "performance.issues.n_plus_one_api_calls.total_duration": 300,
+                "performance.issues.consecutive_http.min_time_saved_threshold": 2000,
             }
         ):
             with self.feature(PERFORMANCE_ISSUE_FEATURES):
@@ -110,6 +111,7 @@ class ProjectPerformanceIssueSettingsTest(APITestCase):
             assert response.data["uncompressed_asset_size_threshold"] == 200000
             assert response.data["consecutive_db_min_time_saved_threshold"] == 300
             assert response.data["n_plus_one_api_calls_total_duration_threshold"] == 300
+            assert response.data["consecutive_http_spans_min_time_saved_threshold"] == 2000
 
             get_value.return_value = {
                 "n_plus_one_db_duration_threshold": 10000,
@@ -122,6 +124,7 @@ class ProjectPerformanceIssueSettingsTest(APITestCase):
                 "file_io_on_main_thread_duration_threshold": 33,
                 "consecutive_db_min_time_saved_threshold": 5000,
                 "n_plus_one_api_calls_total_duration_threshold": 500,
+                "consecutive_http_spans_min_time_saved_threshold": 1000,
             }
 
             with self.feature(PERFORMANCE_ISSUE_FEATURES):
@@ -140,6 +143,7 @@ class ProjectPerformanceIssueSettingsTest(APITestCase):
             assert response.data["file_io_on_main_thread_duration_threshold"] == 33
             assert response.data["consecutive_db_min_time_saved_threshold"] == 5000
             assert response.data["n_plus_one_api_calls_total_duration_threshold"] == 500
+            assert response.data["consecutive_http_spans_min_time_saved_threshold"] == 1000
 
     def test_get_returns_error_without_feature_enabled(self):
         with self.feature({}):
