+++ tests/sentry/utils/performance_issues/test_performance_detection.py
@@ -182,7 +182,7 @@ class PerformanceDetectionTest(TestCase):
 
         default_settings = get_detection_settings(self.project)
 
-        assert default_settings[DetectorType.N_PLUS_ONE_DB_QUERIES]["duration_threshold"] == 100
+        assert default_settings[DetectorType.N_PLUS_ONE_DB_QUERIES]["duration_threshold"] == 90
         assert (
             default_settings[DetectorType.RENDER_BLOCKING_ASSET_SPAN]["fcp_ratio_threshold"] == 0.33
         )
@@ -193,9 +193,10 @@ class PerformanceDetectionTest(TestCase):
         )
         assert default_settings[DetectorType.UNCOMPRESSED_ASSETS]["duration_threshold"] == 300
         assert default_settings[DetectorType.CONSECUTIVE_DB_OP]["min_time_saved"] == 100
-        assert default_settings[DetectorType.SLOW_DB_QUERY][0]["duration_threshold"] == 1000
         assert default_settings[DetectorType.N_PLUS_ONE_API_CALLS]["total_duration"] == 300
         assert default_settings[DetectorType.LARGE_HTTP_PAYLOAD]["payload_size_threshold"] == 300000
+        assert default_settings[DetectorType.CONSECUTIVE_HTTP_OP]["min_time_saved"] == 2000
+        assert default_settings[DetectorType.SLOW_DB_QUERY][0]["duration_threshold"] == 900
 
         self.project_option_mock.return_value = {
             "n_plus_one_db_duration_threshold": 100000,
@@ -208,6 +209,7 @@ class PerformanceDetectionTest(TestCase):
             "file_io_on_main_thread_duration_threshold": 33,
             "consecutive_db_min_time_saved_threshold": 500,
             "n_plus_one_api_calls_total_duration_threshold": 500,
+            "consecutive_http_spans_min_time_saved_threshold": 1000,
         }
 
         configured_settings = get_detection_settings(self.project)
@@ -232,6 +234,7 @@ class PerformanceDetectionTest(TestCase):
         assert configured_settings[DetectorType.DB_MAIN_THREAD][0]["duration_threshold"] == 50
         assert configured_settings[DetectorType.FILE_IO_MAIN_THREAD][0]["duration_threshold"] == 33
         assert configured_settings[DetectorType.CONSECUTIVE_DB_OP]["min_time_saved"] == 500
+        assert configured_settings[DetectorType.CONSECUTIVE_HTTP_OP]["min_time_saved"] == 1000
 
     @override_options(BASE_DETECTOR_OPTIONS)
     def test_n_plus_one_extended_detection_no_parent_span(self):
